+++ aimbots-src/src/utils/nxp_imu/magnetometer/ist8310.hpp
